created testhelpers to reuse test functions
